// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/reportDetailAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { EntityUserMappingAC } from './entityUserMappingAC';
import { KeyValuePairOfIntegerAndString } from './keyValuePairOfIntegerAndString';
import { ObservationCategoryAC } from './observationCategoryAC';
import { RatingAC } from './ratingAC';
import { ReportObservationAC } from './reportObservationAC';
import { ReportUserMappingAC } from './reportUserMappingAC';

export interface ReportDetailAC extends BaseModelAC { 
    processId?: string;
    subProcessId?: string;
    entityId?: string;
    reportId?: string;
    reportObservationList?: Array<ReportObservationAC>;
    ratingList?: Array<RatingAC>;
    observationTypeList?: Array<KeyValuePairOfIntegerAndString>;
    disposition?: Array<KeyValuePairOfIntegerAndString>;
    observationStatus?: Array<KeyValuePairOfIntegerAndString>;
    observationCategoryList?: Array<ObservationCategoryAC>;
    totalReportObservation: number;
    auditorList?: Array<EntityUserMappingAC>;
    responsibleUserList?: Array<EntityUserMappingAC>;
    observationReviewerList?: Array<ReportUserMappingAC>;
    linkedObservationList?: Array<ReportObservationAC>;
}